:truck: move test from example with flutter to pure dart
